+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Replace CSV persistence with JSON persistence
 - Import OE CSV format with custom course fields
 - Export heats startlist as custom OE CSV format
+- Rename Live mode as Finish mode